requirements.txt
python-telegram-bot==20.0
pytelegrambotapi==4.14.0
requests==2.31.0